+++ app/src/main/java/woowacourse/shopping/di/annotation/InjectField.kt
@@ -1,6 +1,5 @@
 package woowacourse.shopping.di.annotation
 
 @Retention(AnnotationRetention.RUNTIME)
-@Target(AnnotationTarget.CONSTRUCTOR)
-annotation class Inject
-
+@Target(AnnotationTarget.FIELD)
+annotation class InjectField
