+++ app/src/main/java/woowacourse/shopping/di/module/DependencyModule.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.di
+package woowacourse.shopping.di.module
 
 import kotlin.reflect.KClass
