+++ app/src/main/java/woowacourse/shopping/global/Application.kt
@@ -1,8 +1,8 @@
 package woowacourse.shopping.global
 
 import android.app.Application
-import woowacourse.shopping.di.DIContainer
-import woowacourse.shopping.di.RepositoryModule
+import woowacourse.shopping.di.core.DIContainer
+import woowacourse.shopping.di.module.RepositoryModule
 
 class Application : Application() {
     override fun onCreate() {
