+++ app/src/test/java/woowacourse/shopping/MainActivityTest.kt
@@ -8,7 +8,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.Robolectric
 import org.robolectric.RobolectricTestRunner
-import woowacourse.shopping.di.DIContainer
+import woowacourse.shopping.di.core.DIContainer
 import woowacourse.shopping.ui.MainActivity
 import woowacourse.shopping.ui.MainViewModel
 import woowacourse.shopping.ui.ViewModelFactory
